The big vore fix
As title states, made vore usable and everything, enjoy!
more specifically, made new damage type to work with bay system, made it recoverable, made belly modes work, fixed audio and ensure content isnt destroyed.

+++ code/modules/mob/typing_indicator.dm
@@ -33,7 +33,7 @@ I IS TYPIN'!'
 	forceMove(get_turf(master))
 
 /mob/proc/create_typing_indicator()
-	if(client && !stat && get_preference_value(/datum/client_preference/show_typing_indicator) == GLOB.PREF_SHOW)
+	if(client && !stat && get_preference_value(/datum/client_preference/show_typing_indicator) == GLOB.PREF_SHOW && !istype(src.loc, /obj/belly))
 		if(typing_indicator)
 			qdel(typing_indicator)
